De-Moo, make max_depth immutable.
As per @leont's remarks about

- Moo being overkill for this context
- Mutability of state being undesirable

This pull fixes both of those issues.

It occurs to me it may be more desirable to have find_modules have a
parameter directly, instead of it being a constructor attribute.

  my $modules = $ml->find_modules('Dist::Zilla', { max_depth => 1 });

Which suffers none of the action-at-a-distance problems requiring code
scanning to determine what a function will do, but also has the utility
of being able to pass a depth parameter on a per-query basis.

( The latter quandry is presently made more difficult by enforcing
  immutability )